Start skeleton of assemblyscript + webpack.
Start skeleton of assemblyscript + webpack.

Javascript doesn't natively support 64-bit ints, only floats. When you attempt to perform bit operations on these numbers, they get cast to 32-bit ints, and so even if you had something in between 32 bits and 53 bits, it would throw away the top 21 bits.

For example, (2**33 + 1) ^ (2**32 + 1) returns 0.

This required way more setup than I imagined it would have, but I'm trying to learn more front-endy stuff so I went with assemblyscript instead of emscripten or something like that. I may switch down the line when I realize that I don't actualyl want to learn front-endy stuff.

The main motivation for wanting bit operations is to support changing all the move computation to bitboards to improve efficiency of the search algorithm.

I also need to do a lot of work on improving the heuristics, but if I did that before moving to bit boards, I'd have to end up re-writing it all anyhow, so I decided to do this first.