Start skeleton for web worker.
Took way longer than I had expected to get working (lol). Web-dev type stuff continues to seem more and more like everything is barely cobbled together, but that may just be becuase there are obvious solutions that I don't know about.

This PR sets up the skeleton for a Web Worker to run the computationally heavy workloads so that we don't block the UI thread.

At the time of this commit, the wasm dateTest call just in the thenable chain should block the dom rendering (so the previous components don't show up, and the checkmark box cannot be checked). Then the dateTest call executed through the web worker does not block the dom.

I now have everything in place to start porting the othello code over to typescript/wasm:
-wp/wasm build chain
-64-bit operations in wasm
-web worker that can run wasm

Once I get the code ported over, and I know a bit more what I'm doing, I ought to clean up the tooling and such, but this is a learning experience so that can wait.