Write to disk.
Interestingly enough, some positions can't be reached from the initial starting empty row.

For example, for:

BBBBBWBW

There is no way to generate the BWBW pattern on positions 4567. No matter the ordering, either position 5 or 6 would have been flipped. This is true regardless of the moves in 0123, and even holds true for real games, since here we're assuming that along this single row, any move in an empty position can be made. Curious.