Huge endgame bug fix :)
When we near end game, and can compute to the end of the table, the code that explicitly says the winner of the match was broken, and it was returning some heuristic value.

Messed this up at some point when i changed it from setHeuristic to getHeuristic.

Now able to beat the renegade Impossible CPU sometimes (though some moves take 30+ seconds).

Added a game that won 56-8!